FINISH the parser

+++ src/meta.rs
@@ -1,12 +1,12 @@
 pub trait Meta {
-    fn size() -> usize;
+    fn size() -> u64;
 
     fn from_bytes(bytes: &[u8]) -> Self;
     fn to_bytes(&self) -> Box<[u8]>;
 }
 
 impl Meta for () {
-    fn size() -> usize {
+    fn size() -> u64 {
         0
     }
